Add files via upload
This is the reviews page so whenever the user changes the rating then its automatically updated in the database. the rating has the step up of 0.5 which means they can rate the app 0.5,1.5,1..... up to 5 stars. they can also write a comment and when they press the comment button then that comment will be inserted in the reviews collection along with the rating(stars).