Add files via upload
This is the page where patients are able to make a review of the app. The rating bar has a step up count to 0.5 which means the user can rate the app from 0.5,1.... to 5 stars rating. They are also able to make a comment and once they press the comment button that comment will be inserted into the reviews collection along with the star rating as well.